+++ src/GIntervalsMeta1D.h
@@ -15,9 +15,9 @@ public:
 
 	struct ChromStat {
 		bool    contains_overlaps;
-		size_t  size;
-		size_t  unified_overlap_size;
-		size_t  unified_touching_size;
+		uint64_t  size;
+		uint64_t  unified_overlap_size;
+		uint64_t  unified_touching_size;
 		int64_t range;
 		int64_t unified_overlap_range;
 
@@ -40,9 +40,9 @@ public:
 
 	//-------------------------------- GIntervalsFetcher1D interface -----------------------------------
 
-	virtual size_t size() const { return m_size; }
+	virtual uint64_t size() const { return m_size; }
 
-	virtual size_t size(int chromid) const { return (*m_user_chrom2size)[chromid]; }
+	virtual uint64_t size(int chromid) const { return (*m_user_chrom2size)[chromid]; }
 
 	virtual int num_chroms() const;
 
@@ -56,7 +56,7 @@ protected:
 	vector<int64_t>            m_chrom2unified_touching_size;
 	vector<int64_t>            m_chrom2range;
 	vector<int64_t>            m_chrom2unified_overlap_range;
-	size_t                     m_size;
+	uint64_t                     m_size;
 	int64_t                    m_range;
 	bool                       m_contains_overlaps;
 	GenomeChromKey            *m_chromkey;
